refactor(sdk): SDK公開ビルドをscripts/からpackages/sdk自身に移設
「CLIとかをscripts/ではなくsdkに含めないの？」という指摘への対応。他パッケージ
（shared/db等）は自分のpackage.jsonにbuildスクリプトを持つ規約だが、SDK公開ビルド
（旧scripts/build-sdk.mjs）だけがrepo-root scripts/に置かれ、packages/sdk自体は
ビルド方法を何も知らない状態だった。将来 ubichill build/lock のようなCLIを同じ
ubichillパッケージに同居させる（bin フィールド追加）ことを考えると、その土台となる
ビルドロジックが repo-only の scripts/ に閉じているのは一貫性が無く、外部から再利用
できない構造だった。

- packages/sdk/build.mjs（旧 scripts/build-sdk.mjs）: root相対パス計算を削除し
  __dirname基準に単純化。出力先デフォルトも packages/sdk/dist-npm/ に変更（他パッケージの
  dist/ と同じくパッケージ自己完結）。
- packages/sdk/build.test.mjs / build.dts.verify.mjs（旧 scripts/build-sdk.test.mjs /
  build-sdk.dts.verify.mjs）: 同様に移設。
- packages/sdk/package.json: "build"/"clean" script と devDependencies
  （dts-bundle-generator, esbuild, rimraf）を自パッケージに持たせる（shared/db と同じ規約）。
  ルートのdts-bundle-generator依存は不要になったため削除（esbuildはbuild-workers.mjs等が
  引き続き使うためルートにも残す）。
- ルート package.json の build:sdk は `pnpm --filter @ubichill/sdk build` に変更。
- vitest.config.ts: include に `packages/*/*.test.mjs`（パッケージルート直下のビルド
  ツールテスト）を追加。vitest.sdk-verify.config.ts の include パスも新しい場所に追従。
- .github/workflows/ci.yml: コメント中の旧パス参照を更新。

検証: pnpm test（213 tests, ~1秒）・pnpm verify:sdk-types（4 tests, ~23秒）・
pnpm typecheck・pnpm lint 全てクリーン。pnpm build:sdk が packages/sdk/dist-npm/ に
正しく出力（npm pack --dry-run で8ファイル確認）。pnpm build:workers は hash不変で
回帰なし。
