+++ package.json
@@ -10,7 +10,7 @@
     "docker:up": "node scripts/start-mods.mjs up -d",
     "docker:down": "node scripts/start-mods.mjs down",
     "build:workers": "node scripts/build-workers.mjs",
-    "build:sdk": "node scripts/build-sdk.mjs",
+    "build:sdk": "pnpm --filter @ubichill/sdk build",
     "gen:lock": "tsx packages/loader/src/cli.ts",
     "verify:mod-locks": "node scripts/verify-mod-locks.mjs",
     "verify:sdk-types": "vitest run --config vitest.sdk-verify.config.ts",
@@ -51,7 +51,6 @@
     "@typescript/native-preview": "7.0.0-dev.20260421.2",
     "concurrently": "^9.2.1",
     "cross-env": "^10.1.0",
-    "dts-bundle-generator": "^9.5.1",
     "esbuild": "^0.28.0",
     "jsdom": "^29.1.1",
     "kill-port": "^2.0.1",
